LayoutTests/fast/multicol/vertical-rl/break-properties-expected.txt

PASS: 'break-before' is at (748, 218)
PASS: 'after-break' is at (748, 428)
